packages 06/05/2022

+++ exercise6/Arithematic/Div.java
@@ -13,7 +13,7 @@ public class Div implements Division{
 		double r=0;
 
 	public void Result(int a,int b){
-		r=a/b;
+		r= (double)a/b;
 		System.out.println("Division Result");
 		System.out.println("Result="+(r)+"\n----------------");
